requirements.txt
# Core app dependencies
streamlit==1.33.0
python-dotenv==1.0.1

# PDF processing
PyMuPDF==1.23.9
PyPDF2==3.0.1
tqdm==4.66.2

# Gemini (Google Generative AI)
google-generativeai==0.5.4

# ElevenLabs Text-to-Speech
elevenlabs==1.1.0
